% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-wrappers.R
\name{dc_combine_bbox}
\alias{dc_combine_bbox}
\title{dc_combine_bbox}
\usage{
dc_combine_bbox(list_layers)
}
\arguments{
\item{list_layers}{A list of sf layers}
}
\value{
The bounding box as a vector of doubles (as sf::st_bbox)
}
\description{
Takes a list of sf layers and compute the bounding box
that covers them all.
}
