% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_concentric_circles.R
\name{dc_concentric_circles}
\alias{dc_concentric_circles}
\title{dc_concentric_circles}
\usage{
dc_concentric_circles(
  positioning_result,
  steps = list(30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330, 360)
)
}
\arguments{
\item{positioning_result}{The object returned by the \code{dc_move_from_reference_point}
function.}

\item{steps}{The steps (in the unit of the durations used ine the
\code{dc_move_from_reference_point} function) at which creating the circles.}
}
\value{
A sf LINESTRING object
}
\description{
Create concentric circles around the reference point, using the results of the
\code{dc_move_from_reference_point}.
}
