% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-wrappers.R
\name{dc_create}
\alias{dc_create}
\title{dc_create}
\usage{
dc_create(source_points, image_points, precision, bbox, niter)
}
\arguments{
\item{source_points}{The source point layer}

\item{image_points}{The image point layer}

\item{precision}{The precision of the grid to be created
(a higher value means a higher precision - 0.5 gives usually a coarse result,
2 is good default, 4 is very detailed).}

\item{bbox}{The bounding box of the grid to be created}

\item{niter}{The number of iterations (using \code{round(4 * sqrt(length(source_points)))}
is a good default for this value)}
}
\value{
An interpolation grid to be used to transform the layers
}
\description{
Create the interpolation grid.
}
