% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_generate_positions_from_durations.R
\name{dc_generate_positions_from_durations}
\alias{dc_generate_positions_from_durations}
\title{dc_generate_positions_from_durations}
\usage{
dc_generate_positions_from_durations(
  durations,
  source_points,
  adjustment_type = "euclidean"
)
}
\arguments{
\item{durations}{The durations matrix}

\item{source_points}{The source points, an sf POINT object}

\item{adjustment_type}{The adjustment type to use, either "euclidean" or "affine"}
}
\value{
A list object with the source points and the image points, ready
to be used with the dc_create function
}
\description{
Generate positions from durations matrix, using PCoA to find the relative
positions between points, then perform a procrustes analysis to find the
best fit between the source points and the image points.
}
