% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-wrappers.R
\name{dc_interpolate}
\alias{dc_interpolate}
\title{dc_interpolate}
\usage{
dc_interpolate(interpolation_grid, layer_to_deform)
}
\arguments{
\item{interpolation_grid}{The interpolation grid}

\item{layer_to_deform}{The sf layer to interpolate}
}
\value{
The sf layer deformed by the interpolation grid
}
\description{
Interpolate a sf layer using the interpolation grid.
}
