% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{dc_interpolate_parallel}
\alias{dc_interpolate_parallel}
\title{dc_interpolate_parallel}
\usage{
dc_interpolate_parallel(interpolation_grid, layers_to_deform)
}
\arguments{
\item{interpolation_grid}{The interpolation grid}

\item{layers_to_deform}{A list of sf layers to interpolate}
}
\value{
The sf layers deformed by the interpolation grid
}
\description{
Interpolate a list of sf layers using the interpolation grid.
}
