% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{dc_move_from_reference_point}
\alias{dc_move_from_reference_point}
\title{dc_move_from_reference_point}
\usage{
dc_move_from_reference_point(
  reference_point,
  other_points,
  duration_col_name,
  factor
)
}
\arguments{
\item{reference_point}{The point from which the other points will be moved,
an sf POINT object}

\item{other_points}{The other points to move, an sf POINT object}

\item{duration_col_name}{The name of the column containing the durations
in the other_points sf object}

\item{factor}{The factor of displacement (default: 1)}
}
\value{
A list object with the source points and the image points, ready
to be used with the dc_create function
}
\description{
Move points from a reference point using durations between the reference point
and all the other points.
}
\examples{
library(sf)
start <- st_read(dsn = system.file("gpkg/pit.gpkg", package = "distanamo"),
                 layer = "start", quiet = TRUE)
points <-  st_read(dsn = system.file("gpkg/pit.gpkg", package = "distanamo"),
                   layer = "points", quiet = TRUE)
center <-  st_read(dsn = system.file("gpkg/pit.gpkg", package = "distanamo"),
                   layer = "center", quiet = TRUE)

pos_result <- dc_move_from_reference_point(
  reference_point = start,
  other_points = points,
  duration_col_name = "durations",
  factor = 1
)

plot(st_geometry(pos_result$source_points), pch = 4, cex = .5)
plot(st_geometry(pos_result$image_points), add = TRUE, pch = 4, cex = .5, col = "red")
plot(st_geometry(center), add = TRUE)
}
