% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{distanamo}
\alias{distanamo-package}
\alias{distanamo}
\title{Create Distance Cartograms}
\description{
This package enables the creation of what is often defined as a distance
cartogram. This is done by showing (on background(s) layer(s), such as the
territorial divisions of the study zone) the local deformations (calculated
using Waldo Tobler's bidimensional regression) to fit image points to
source points.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/riatelab/distanamo/issues/}
}

}
\author{
\strong{Maintainer}: Matthieu Viry \email{matthieu.viry@cnrs.fr} (\href{https://orcid.org/0000-0002-0693-8556}{ORCID})

Other contributors:
\itemize{
  \item Timothée Giraud \email{timothee.giraud@cnrs.fr} (\href{https://orcid.org/0000-0002-1932-3323}{ORCID}) [contributor]
}

}
