% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.adjustment_result.R
\name{plot.adjustment_result}
\alias{plot.adjustment_result}
\title{plot adjustment_result object}
\usage{
\method{plot}{adjustment_result}(x, ...)
}
\arguments{
\item{x}{object of class adjustment_result}

\item{...}{further specifications, see \link{plot} for details}
}
\description{
Plot the result of dc_adjust
}
\examples{
library(sf)

# Read source points
source_pts <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "prefecture", quiet = TRUE
)

# Read non adjusted image points
image_pts_not_adj <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "image-points-not-adjusted", quiet = TRUE
)

# Read the background layer to deform
background_layer <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "departement", quiet = TRUE
)

# Adjust image points to source points
adj_result <- dc_adjust(
  source_points = source_pts,
  image_points = image_pts_not_adj,
  "euclidean"
)

# Plot result of the adjustment step
plot(adj_result)

# Summary statistics of the adjustment step
summary(adj_result)

}
