% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.interpolation_grid.R
\name{plot.interpolation_grid}
\alias{plot.interpolation_grid}
\title{plot interpolation_grid object}
\usage{
\method{plot}{interpolation_grid}(
  x,
  which = 1:4,
  ask = interactive(),
  caption = list("Source grid", "Interpolated grid", "Image to interpolated points",
    "Deformation strength"),
  ...
)
}
\arguments{
\item{x}{object of class interpolation_grid}

\item{...}{further specifications, see \link{plot} for details}
}
\description{
Plot the interpolation grid, resulting from dc_create
}
\examples{
library(sf)

# Read source points
source_pts <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "prefecture", quiet = TRUE
)

# Read image points
image_pts <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "image-points", quiet = TRUE
)

# Read the background layer to deform
background_layer <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "departement", quiet = TRUE
)

# Create the interpolation grid
igrid <- dc_create(
  source_points = source_pts,
  image_points = image_pts,
  precision = 2,
  bbox = st_bbox(background_layer)
)

# Plot various depictions of the interpolation grid
plot(igrid)

# Useful information about the interpolation grid
summary(igrid)

}
