% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.interpolation_grid.R
\name{summary.interpolation_grid}
\alias{summary.interpolation_grid}
\title{summary interpolation_grid object}
\usage{
\method{summary}{interpolation_grid}(object, ...)
}
\arguments{
\item{object}{object of class interpolation_grid}

\item{...}{further specifications, see \link{summary} for details}
}
\description{
Compute summary statistics for interpolation_grid
}
\examples{
library(sf)

# Read source points
source_pts <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "prefecture", quiet = TRUE
)

# Read image points
image_pts <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "image-points", quiet = TRUE
)

# Read the background layer to deform
background_layer <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "departement", quiet = TRUE
)

# Create the interpolation grid
igrid <- dc_create(
  source_points = source_pts,
  image_points = image_pts,
  precision = 2,
  bbox = st_bbox(background_layer)
)

# Plot various depictions of the interpolation grid
plot(igrid)

# Useful information about the interpolation grid
summary(igrid)

}
