% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_adjust.R
\name{dc_adjust}
\alias{dc_adjust}
\title{dc_adjust}
\usage{
dc_adjust(source_points, image_points, adjustment_type = "euclidean")
}
\arguments{
\item{source_points}{The source point layer, sf POINT object}

\item{image_points}{The layer of point to be adjusted to fit \code{source_points},
sf POINT object}

\item{adjustment_type}{The adjustment type to use, either "euclidean" or "affine"}
}
\value{
A list object with the transformation matrix, various metrics and
the adjusted points
}
\description{
Computes the positions of the adjusted points by fitting image points to source points.
}
