% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_combine_bbox.R
\name{dc_combine_bbox}
\alias{dc_combine_bbox}
\title{dc_combine_bbox}
\usage{
dc_combine_bbox(list_layers)
}
\arguments{
\item{list_layers}{A list of sf objects}
}
\value{
An sf bounding box is returned.
}
\description{
Takes a list of sf objects and compute the bounding box
that covers them all.
}
\examples{
library(sf)

# Read source points
source_pts <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "prefecture", quiet = TRUE
)

# Read the background layer to deform
background_layer <- st_read(
  dsn = system.file("gpkg/data-prefecture.gpkg", package = "distanamo"),
  layer = "departement", quiet = TRUE
)

bbox <- dc_combine_bbox(list(source_pts, background_layer))
bbox
}
