% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{dc_generate_positions_from_durations}
\alias{dc_generate_positions_from_durations}
\title{dc_generate_positions_from_durations}
\usage{
dc_generate_positions_from_durations(durations, source_points)
}
\arguments{
\item{durations}{The durations matrix}

\item{source_points}{The source points, an sf POINT object}
}
\value{
An sf POINT object is returned.
}
\description{
Generate positions from durations matrix.
}
