% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{dc_generate_positions_from_durations}
\alias{dc_generate_positions_from_durations}
\title{dc_generate_positions_from_durations}
\usage{
dc_generate_positions_from_durations(durations, source_points)
}
\arguments{
\item{durations}{The durations matrix}

\item{source_points}{The source points, an sf POINT object}
}
\value{
A list object with the source points and the image points, ready
to be used with the dc_create function
}
\description{
Generate positions from durations matrix, using PCoA to find the relative
positions between points, then perform a procrustes analysis to find the
best fit between the source points and the image points.
}
