% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{dc_move_points}
\alias{dc_move_points}
\title{dc_move_points}
\usage{
dc_move_points(points, times, factor)
}
\arguments{
\item{points}{The points to be moved}

\item{times}{The times between the points (note that)
the reference point should have a time of 0 and that times
and points should be in the same order.}

\item{factor}{The factor of displacement (default: 1)}
}
\value{
The moved points
}
\description{
Move points.
}
