% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{distanamo}
\alias{distanamo-package}
\alias{distanamo}
\title{Create Distance Cartograms}
\description{
This package enables the creation of what is often defined as a distance
cartogram. Distance cartograms are a type of cartogram that deforms the
layers of a map according to the distancesbetween a set of source points
and a set of image points.
This is done by extending (by interpolation) to the layer(s) of the study
area (territorial divisions, network...) the local displacement between the
source coordinates and the image coordinates, derived from the distances
between each pair of homologous points (source / image points).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/riatelab/distanamo/issues/}
}

}
\author{
\strong{Maintainer}: Matthieu Viry \email{matthieu.viry@cnrs.fr} (\href{https://orcid.org/0000-0002-0693-8556}{ORCID})

Other contributors:
\itemize{
  \item Timothée Giraud \email{timothee.giraud@cnrs.fr} (\href{https://orcid.org/0000-0002-1932-3323}{ORCID}) [contributor]
}

}
